+++ CHANGES.rst
@@ -2,7 +2,7 @@ ChangeLog
 =========
 
 
-0.4.0 (unreleased)
+0.4.0 (2015-03-25)
 ------------------
 
 * Show subdivision as-is in rendered address if not already printed by state,